Move tests to test_model_sets
